+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ---
 
 ## [Unreleased]
+---
 ## [2.0.1-beta] - 2025-08-04
 ### Added
 - "/" endpoint to api which returns api name, version, and status.
@@ -17,7 +18,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - backend bug fix, reset scan cache counter after each iteration.
 - backend changed login endpoint behaviour. On valid id, data is restored from databased and loaded to cache.
 - fix `<label for=""> `usage in LocationRadio.vue and ShiftSizeRadio.vue to match RadioButton "inputId".
----
 ## [2.0.0-beta] - 2025-08-04
 
 ### Changed